+++ website/messages.js
@@ -1,0 +1,197 @@
+window.messages = [
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 4, 10),
+		text: 'Looks to me like it’s going to be SleepyCreepy Joe over Crazy Bernie. Everyone else is fading fast!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 4, 10),
+		text: 'As part of the @ObamaFoundation, moms like Dr. Mwansa Ketty Lubeya are answering the call for change—and raising the next generation of leaders while they’re at it.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 4, 10),
+		text: 'Republicans must stick together!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 4, 1),
+		text: 'Happy #CollegeSigningDay, everybody! I\'m proud of every student out there who\'s taking the next step in their education. It\'s the best thing you can do for yourself and your future.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 4, 1),
+		text: 'Great Consumer Price Index just out. Really good, very low inflation! We have a great chance to “really rock!” Good numbers all around.'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 30),
+		text: 'Condolences to the family of John Singleton. His seminal work, Boyz n the Hood, remains one of the most searing, loving portrayals of the challenges facing inner-city youth. He opened doors for filmmakers of color to tell powerful stories that have been too often ignored.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 30),
+		text: 'Your all time favorite President got tired of waiting for China to help out and start buying from our FARMERS, the greatest anywhere in the World!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 30),
+		text: 'Congratulations to the new class of @ObamaFoundation Fellows! Their work in every region of the globe is already making the world better, and I hope this experience helps them make an even bigger impact. Obama.org/fellowship'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 29),
+		text: 'Build your products in the United States and there are NO TARIFFS!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 28),
+		text: 'For thirty-six years, Richard Lugar proved that pragmatism and decency work—not only in Washington but all over the world. Michelle and I send our warmest sympathies to his family and all those who had the privilege of knowing this American statesman.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 26),
+		text: 'The average 401(k) balance has SOARED since the bottom of the market - 466%. Wow!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 24),
+		text: 'Dominique’s story is a reminder to us that where you’re from shouldn’t define what you can achieve. I’m proud of her and all students striving for a better future.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 23),
+		text: 'V.P. Mike Pence will be interviewed on @foxandfriends at 7:30 A.M. Enjoy!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 22),
+		text: 'This generation of climate activists is tired of inaction, and they\'ve caught the attention of leaders all over the world. So while this challenge is only getting more urgent, they show us the kind of action it\'ll take to meet this moment.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 22),
+		text: 'We have lost 500 Billion Dollars a year, for many years, on Crazy Trade with China. NO MORE!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 22),
+		text: 'And in the U.S., 13-year old @AlexandriaV2005 of New York, 13-year old @havenruthie of Denver, and 16-year old @israhirsi of Minneapolis are some of the young people leading the fight at home. I hope more of us will follow their example.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 22),
+		text: 'Today, it was my honor to welcome the 2018 World Series Champion Boston @RedSox to the @WhiteHouse!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 22),
+		text: 'I met 23-year-old @luisamneubauer in Berlin earlier this month. Luisa, the organizer behind the student movement in Germany, says, “It’s our future and we’re not going to stop fighting for it.”'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 22),
+		text: 'MAKE AMERICA GREAT AGAIN!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 22),
+		text: 'They\'re people like 16-year-old @GretaThunberg, whose protests at Swedish parliament sparked a movement. Inspired by Greta\'s action, Fridays for Future brought together more than a million strikers on every continent last month to demand action on climate.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 22),
+		text: 'Beautiful evening in Panama City Beach, Florida. Thank you! #MAGA'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 22),
+		text: 'Young people all over the world are leading the way in the fight to protect our planet because they know their future depends on it. This Earth Day, I want to celebrate the courageous, committed young leaders who are stepping up to save the one planet we\'ve got.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 21),
+		text: 'Big crowds in Panama City Beach, Florida. See everyone in 30 minutes! @FoxNews @TuckerCarlson @SeanHannity'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 21),
+		text: 'To all who celebrate today, happy Easter from our family to yours! On this day of rebirth and renewal, let’s recommit to love and serve our brothers and sisters, especially those in need, in every way we can. Have a great day, everybody.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 21),
+		text: 'Thank you @NewtGingrich & @FoxandFriends!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 21),
+		text: 'The attacks on tourists and Easter worshippers in Sri Lanka are an attack on humanity. On a day devoted to love, redemption, and renewal, we pray for the victims and stand with the people of Sri Lanka.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 17),
+		text: '“CASE CLOSED!” @SenateMajLDR'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 15),
+		text: 'Notre Dame is one of the world’s great treasures, and we’re thinking of the people of France in your time of grief. It’s in our nature to mourn when we see history lost – but it’s also in our nature to rebuild for tomorrow, as strong as we can.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 14),
+		text: '“This British Spy, Christopher Steele, tried so hard to get this (the Fake Dossier) out before the Election. Why?” @kilmeade @foxandfriends'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 13),
+		text: 'Another good story worth sharing: From one "kid from Akron" to a new generation of Akron kids, some remarkable early achievements at @IPROMISESchool. Great work, @KingJames — and even better work by those students. Proud to be a witness to their success:'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 9),
+		text: 'Big Court win at our Southern Border! We are getting there - and Wall is being built!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 6),
+		text: 'In just a few minutes, I’m taking the stage at the @ObamaFoundation Town Hall Europe for a conversation about how to best support civic leaders carrying out good work across the continent. I hope you’ll tune in live at (link: http://obama.org) obama.org.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 5),
+		text: '“The real “Obstruction of Justice” is what the Democrats are trying to do to this Attorney General.” Congressman Jim Jordan (R-Ohio). @MariaBartiromo'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 4),
+		text: 'Here’s a story about people doing good that\'s worth sharing: @BradPaisley and his wife @Kimwilliamspais saw food insecurity in Nashville and decided to do something about it. They just broke ground on a free grocery store to help families in need:'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 3),
+		text: 'I am pleased to inform you that THE BIG FIREWORKS, after many years of not having any, are coming back to beautiful Mount Rushmore in South Dakota. Great work @GovKristiNoem and @SecBernhardt! #MAGA'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 3),
+		text: 'Great to see Chicago’s historic mayoral race between two highly qualified candidates. Congrats to our next mayor, Lori Lightfoot—and Toni Preckwinkle campaigned hard and did us proud. I know that with our city’s heart and Lori’s leadership, Chicago’s best days are still ahead.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 2),
+		text: 'He wants to impeach because they can’t win election. Sad!'
+	},
+	{
+		username: 'barackobama',
+		date: new Date(2019, 3, 2),
+		text: 'Valerie is one of my oldest friends and advisors – she was by my side when I first decided to run for office and for every major moment of the presidency. @ValerieJarrett\'s voice has often inspired me and I know “Finding My Voice” will inspire others to lift their voices, too.'
+	},
+	{
+		username: 'realdonaldtrump',
+		date: new Date(2019, 3, 2),
+		text: '"Forgotten Man" Story: Under Trump, Red Counties Economically Thrive (link) via @BreitbartNews'
+	}
+]